pyproject.toml
[tool.poetry]
name = "gitbrew"
version = "0.1.0"
description = "LLM-powered github utility"
authors = ["Navneet Desai <navneetdesai44@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pre-commit = "^3.3.3"
black = "^23.7.0"
isort = "^5.12.0"
pytest = "^7.4.0"
openai = "^0.28.0"
langchain = "^0.0.295"
pygithub = "^1.59.1"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.1"
chromadb = "^0.4.11"
bs4 = "^0.0.1"
cohere = "^4.27"
pyinquirer = "^1.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
